unified audio check with expectTotalAudioEnergyToBeGreaterThan
